mqtt_client/requirements.txt
Package            Version
------------------ --------
certifi            2024.7.4
charset-normalizer 3.3.2
gitdb              4.0.11
GitPython          3.1.41
idna               3.7
linecache2         1.0.0
paho-mqtt          1.5.1
pip                24.0
requests           2.32.3
setuptools         69.0.3
six                1.16.0
smmap              5.0.1
traceback2         1.4.0
unittest2          1.1.0
urllib3            2.2.2
wheel              0.43.0
